+++ lib/layers/adapters.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'layers/adapters/pagination/kaminari'
+require 'layers/adapters/pagination/will_paginate'
+require 'layers/adapters/relation/active_record'
+require 'layers/adapters/relation/duck_type'
+
+module Layers
+  module Adapters
+  end
+end