+++ lib/layers/errors.rb
@@ -2,4 +2,5 @@
 
 module Layers
   class Error < StandardError; end
+  class ConfigurationError < Layers::Error; end
 end